update: enable ap, add restart button, use ALWAYS_ON
the switch reports very weak WiFi signal all the time (-90db). enable AP
so that it can be managed remotely even when the switch cannot associate
to the AP.

the real fix is to move the switch to a better location, or install
another WiFi AP at the back garden.